requirements.txt
# Core dependencies (updated for security)
numpy>=1.22.0
pandas>=1.3.0
scikit-learn>=1.5.0
joblib>=1.2.0
xgboost>=1.7.0

# Data and API (updated for security)
requests>=2.32.4  # Used for FMP, Alpha Vantage, and other API calls
redis>=4.5.3
python-dotenv>=0.19.0

# Database
psycopg2-binary>=2.9.0

# Security
cryptography>=3.4.8

# Real-time data
websocket-client>=1.3.0

# Monitoring (updated for security)
prometheus-client>=0.12.0
flask>=2.3.0

# Machine Learning and RL
tensorflow>=2.13.0
optuna>=3.0.0
psutil>=5.9.0

# Logging and utilities
colorama>=0.4.4
